backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==24.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
Faker==22.0.0

